+++ NEWS.md
@@ -1,4 +1,13 @@
 
+# vtreat 1.3.0 2018/07/20
+
+ * Fix z-split annotation issue.
+ * Change designTreatmentsZ() defaults.
+ * Documentation fixes.
+ * Multiclass treatmetns vignette.
+ * Simple treatments.
+ * Fix CRAN Note on pre 3.5.0 R parallel usage.
+
 # vtreat 1.2.3 2018/07/11
 
   * Smooth catP novel levels scoring a bit.